--- core/src/com/geometric/wars/cube/Mountable.java
@@ -1,7 +1,0 @@
-package com.geometric.wars.cube;
-
-
-public interface Mountable {
-    public MountableView getView();
-}
-
