--- core/src/com/geometric/wars/cube/MountableView.java
@@ -1,21 +1,0 @@
-package com.geometric.wars.cube;
-
-import com.badlogic.gdx.graphics.g2d.TextureRegion;
-import com.badlogic.gdx.graphics.g3d.Material;
-import com.badlogic.gdx.graphics.g3d.model.Node;
-import com.badlogic.gdx.graphics.g3d.utils.ModelBuilder;
-import com.badlogic.gdx.math.Vector3;
-import com.geometric.wars.Direction3D;
-
-public interface MountableView {
-    Material getMaterial();
-
-    Node getNode();
-
-    /**
-     * should contain rules to build part of model
-     **/
-    void buildMeshPart(ModelBuilder modelBuilder);
-    Vector3 getSize();
-}
-
