+++ core/src/com/geometric/wars/cube/mountables/Mountable.java
@@ -1,0 +1,7 @@
+package com.geometric.wars.cube.mountables;
+
+
+public interface Mountable {
+    MountableView getView();
+}
+
