+++ core/src/com/geometric/wars/cube/mountables/MountableGun.java
@@ -1,0 +1,5 @@
+package com.geometric.wars.cube.mountables;
+
+public interface MountableGun extends Mountable {
+    //TODO add gun specific parameters
+}
